+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * [#3164](https://github.com/bbatsov/rubocop/pull/3164): Add [Fastlane](https://fastlane.tools/)'s Fastfile to the default Includes. ([@jules2689][])
 * [#3173](https://github.com/bbatsov/rubocop/pull/3173): Make `Style/ModuleFunction` configurable with `module_function` and `extend_self` styles. ([@tjwp][])
 * [#3105](https://github.com/bbatsov/rubocop/issues/3105): Add new `Style/RequestReferer` cop. ([@giannileggio][])
+* [#3188](https://github.com/bbatsov/rubocop/issues/3188): Add new `Style/RaiseString` cop. ([@tjwp][])
 
 ### Bug fixes